@@ -21,14 +21,15 @@
         }
     ],
     "require": {
-        "php": "^7.4|^8.0|^8.1",
-        "spatie/laravel-medialibrary": "^8.0|^9.0|^10.0|^11.0"
+        "php": "^8.2",
+        "spatie/laravel-medialibrary": "^11.0",
+        "spatie/laravel-data": "^3.0"
     },
     "require-dev": {
         "laravel/nova": "^4.0",
-        "orchestra/testbench": "^5.0|^6.0|^7.0",
+        "orchestra/testbench": "^6.0|^7.0|^8.0",
         "nunomaduro/collision": "^4.0|^5.0|^6.0",
-        "phpunit/phpunit": "^8.5|^9.0"
+        "phpunit/phpunit": "^9.0"
     },
     "autoload": {
         "psr-4": {
