@@ -6,30 +6,39 @@ use DmitryBubyakin\NovaMedialibraryField\Fields\Medialibrary;
 use DmitryBubyakin\NovaMedialibraryField\Integrations\NovaDependencyContainer\ResolveFromDependencyContainerFields;
 use DmitryBubyakin\NovaMedialibraryField\Integrations\NovaFlexibleContent\ResolveFromFlexibleLayoutFields;
 use Exception;
+use Laravel\Nova\Fields\FieldCollection;
 use Laravel\Nova\Http\Requests\NovaRequest;
 use Laravel\Nova\Resource;
 
 class MedialibraryFieldResolver
 {
-    public static $resolvers = [
+    public static array $resolvers = [
         ResolveFromFlexibleLayoutFields::class,
         ResolveFromDependencyContainerFields::class,
     ];
 
-    private $request;
-    private $resource;
-    private $attribute;
+    private NovaRequest $request;
 
-    public function __construct(NovaRequest $request, Resource $resource = null, string $attribute = null)
+    private ?Resource $resource;
+
+    private ?string $attribute;
+
+    public function __construct(
+        NovaRequest $request,
+        Resource $resource = null,
+        string $attribute = null,
+    )
     {
         $this->request = $request;
         $this->resource = $resource ?: $request->newResource();
-        $this->attribute = $attribute ?: $request->field;
+        $this->attribute = $attribute ?: $request->route('field');
     }
 
     public function __invoke(): Medialibrary
     {
-        $fields = $this->resource->availableFields($this->request);
+        $fields = $this
+            ->resource
+            ->availableFields($this->request);
 
         foreach (static::$resolvers as $className) {
             $resolver = new $className;
@@ -39,10 +48,13 @@ class MedialibraryFieldResolver
             }
         }
 
-        return $fields
-            ->whereInstanceOf(Medialibrary::class)
-            ->findFieldByAttribute($this->attribute, function (): void {
-                throw new Exception("Field with attribute `{$this->attribute}` is not found.");
-            });
+        /** @var FieldCollection $fields */
+        $fields = $fields->whereInstanceOf(Medialibrary::class);
+
+        $callback = function (): void {
+            throw new Exception("Field with attribute `$this->attribute` is not found.");
+        };
+
+        return $fields->findFieldByAttribute($this->attribute, $callback);
     }
 }
