@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace DmitryBubyakin\NovaMedialibraryField\Fields;
 
+use Illuminate\Support\Collection;
 use Laravel\Nova\Fields\Field;
 use Spatie\MediaLibrary\MediaCollections\Models\Media;
 
@@ -13,17 +14,23 @@ class GeneratedConversions extends Field
 
     public $showOnIndex = false;
 
-    public function __construct($name)
+    public function __construct(string $name)
     {
-        parent::__construct($name, function (Media $media) {
+        $callback = function (Media $media): Collection {
+            $mapCallback = function (string $conversionName) use ($media): array {
+                return [
+                    $conversionName => $media->getFullUrl($conversionName),
+                ];
+            };
+
             return $media
                 ->getGeneratedConversions()
                 ->filter()
                 ->keys()
-                ->mapWithKeys(function (string $conversionName) use ($media) {
-                    return [$conversionName => $media->getFullUrl($conversionName)];
-                });
-        });
+                ->mapWithKeys($mapCallback);
+        };
+
+        parent::__construct($name, $callback);
     }
 
     public function withTooltips(bool $withTooltips = true): self
