@@ -2,7 +2,7 @@
 
 namespace DmitryBubyakin\NovaMedialibraryField\Fields\Support;
 
-use DmitryBubyakin\NovaMedialibraryField\TransientModel;
+use DmitryBubyakin\NovaMedialibraryField\Models\TransientModel;
 use Illuminate\Http\UploadedFile;
 use Spatie\MediaLibrary\HasMedia;
 use Spatie\MediaLibrary\MediaCollections\Models\Media;
@@ -14,7 +14,7 @@ class AttachCallback
         UploadedFile $file,
         string $collectionName,
         string $diskName,
-        string $fieldUuid
+        ?string $fieldUuid
     ): Media {
         // you can override this behaviour: Medialibrary::attachUsing()
         if ($model instanceof TransientModel) {
