@@ -1,33 +1,47 @@
 <?php declare(strict_types=1);
 
-namespace DmitryBubyakin\NovaMedialibraryField;
+namespace DmitryBubyakin\NovaMedialibraryField\Models;
 
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\MorphMany;
 use Illuminate\Database\Query\Builder;
+use Illuminate\Support\Collection;
 use Spatie\MediaLibrary\HasMedia;
 use Spatie\MediaLibrary\InteractsWithMedia;
 use Spatie\MediaLibrary\MediaCollections\Models\Media;
 
+/**
+ * @method static self make()
+ */
 class TransientModel extends Model implements HasMedia
 {
     use InteractsWithMedia;
 
-    public static $customPropertyName = '__target_props__';
+    public static string $customPropertyName = '__target_props__';
 
     public $timestamps = false;
 
     protected $guarded = [];
 
-    public function newInstance($attributes = [], $exists = false)
+    public function newInstance($attributes = [], $exists = false): self
     {
-        return parent::newInstance(TransientModel::instanceAttributes() + $attributes, true);
+        $attributes = array_merge(
+            TransientModel::instanceAttributes(),
+            $attributes,
+        );
+
+        return parent::newInstance($attributes, true);
     }
 
-    public function newBaseQueryBuilder()
+    public static function instanceAttributes(): array
+    {
+        return ['id' => '1'];
+    }
+
+    public function newBaseQueryBuilder(): Builder
     {
         return new class($this->getConnection()) extends Builder {
-            public function get($columns = ['*'])
+            public function get($columns = ['*']): Collection
             {
                 return collect([(object) TransientModel::instanceAttributes()]);
             }
@@ -36,12 +50,9 @@ class TransientModel extends Model implements HasMedia
 
     public function staleMedia(): MorphMany
     {
-        return $this->media()->where('created_at', '<=', now()->subDays(1));
-    }
-
-    public static function instanceAttributes(): array
-    {
-        return ['id' => '1'];
+        return $this
+            ->media()
+            ->where('created_at', '<=', now()->subDay());
     }
 
     public static function setCustomPropertyName(string $name): void
@@ -60,18 +71,6 @@ class TransientModel extends Model implements HasMedia
             'target' => $target,
             'collectionName' => $collectionName,
         ]);
-    }
-
-    public static function getCustomPropertyValue(?Media $media): array
-    {
-        if (
-            is_null($media) ||
-            is_null($value = $media->getCustomProperty(static::getCustomPropertyName()))
-        ) {
-            return [null, null];
-        }
-
-        return [$value['target'], $value['collectionName']];
     }
 
     public function registerAllMediaConversions(Media $media = null): void
@@ -102,5 +101,19 @@ class TransientModel extends Model implements HasMedia
 
             $conversion->nonQueued();
         }
+    }
+
+    public static function getCustomPropertyValue(?Media $media): array
+    {
+        $value = $media?->getCustomProperty(static::getCustomPropertyName());
+
+        if (is_null($media) || is_null($value)) {
+            return [null, null];
+        }
+
+        return [
+            $value['target'],
+            $value['collectionName'],
+        ];
     }
 }
